requirements.txt
aiofiles==22.1.0
altgraph @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
asgiref==3.5.2
colorama==0.4.6
dict-toolbox==2.3.1
dj-database-url==1.2.0
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
future @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
gunicorn==20.1.0
install==1.3.5
Jinja2==3.1.2
lazy-object-proxy==1.8.0
macholib @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==2.1.1
msgpack==1.0.4
nest-asyncio==1.5.6
pop==23.0.0
pop-config==10.2.3
pop-loop==1.0.6
psycopg2-binary==2.9.5
PyJWT==2.8.0
pytz==2022.7
PyYAML==6.0
rend==6.5.1
rest-framework-simplejwt==0.0.2
six @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
sniffio==1.3.0
sqlparse==0.4.3
toml==0.10.2
typing_extensions==4.11.0
whitenoise==6.2.0
